pyproject.toml
[project]
name = "sitgrep"
version = "3.6.8"
description = "Enhance your code review with Sitgrep"
authors = [
    { name = "John Ascher", email = "jascher@securityinnovation.com" }
]
maintainers = [
  { name = "John Ascher", email = "jascher@securityinnovation.com" }
]
readme = "README.md"
dependencies = [
    "semgrep>=1.46.0,<=1.100.0",
    "gitpython>=3.1.43",
    "packaging>=24.0",
    "PyYAML",
    "rich"
]
requires-python = ">=3.7.16"

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
sitgrep = "sitgrep:cli"
